completed course building

+++ Server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section=require("../models/Section");
 const Course=require("../models/Course");
+const SubSection = require("../models/SubSection")
 
 exports.createSection=async(req,res)=>{
     try {
@@ -47,7 +48,7 @@ exports.createSection=async(req,res)=>{
 exports.updateSection=async(req,res)=>{
     try {
         // data fetch
-        const {sectionName,sectionId}=req.body;
+        const {sectionName,sectionId,courseId}=req.body;
         // validation
         if(!sectionId || !sectionName){
             return res.status(400).json({
@@ -57,11 +58,19 @@ exports.updateSection=async(req,res)=>{
         }
         // update data
         const section=await Section.findByIdAndUpdate({_id:sectionId},{sectionName:sectionName},{new:true});
+        const course = await Course.findById(courseId)
+		.populate({
+			path:"courseContent",
+			populate:{
+				path:"subSection",
+			},
+		})
+		.exec();
         // return res
         return res.status(200).json({
             success:true,
             message:"Section updated successfully",
-            section:section,
+            data:course,
         })
     } catch (error) {
         return res.status(500).json({
@@ -75,16 +84,35 @@ exports.updateSection=async(req,res)=>{
 exports.deleteSection=async(req,res)=>{
     try {
         const {courseId,sectionId}=req.body;
-        await Section.findByIdAndDelete(sectionId);
+        const section = await Section.findById(sectionId);
         // HW to update in course
         await Course.findByIdAndUpdate(courseId,
                         {
                             $pull:{courseContent:sectionId,}
                         },
                         );
+        if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+        await SubSection.deleteMany({_id: {$in: section.subSection}});
+
+		await Section.findByIdAndDelete(sectionId);
+
+		//find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
         return res.status(200).json({
             success:true,
             message:"Section deleted successfully",
+            data:course,
         })
     } catch (error) {
         return res.status(500).json({
